protecting the strings.h with an ifdef, and casting isspace() family of macros to match man page (int).

+++ src/geo-find.c
@@ -1,5 +1,5 @@
 
-// $Id: geo-find.c,v 1.9 2005/08/17 19:08:44 we7u Exp $
+// $Id: geo-find.c,v 1.10 2005/08/30 20:28:50 kd6zwr Exp $
 
 
 /* Copyright 2002 Daniel Egnor.  See LICENSE.geocoder file.
@@ -312,7 +312,7 @@ static const char *input_word(struct state *s,const char *pos) {
 	}
 
 	while (s->input_begin != s->input_end 
-	   && (s->input_depth > 0 || !isalnum(*s->input_begin))) {
+	   && (s->input_depth > 0 || !isalnum((int)*s->input_begin))) {
 		const char ch = *s->input_begin++;
 		if ('(' == ch)
 			++s->input_depth;
@@ -322,9 +322,9 @@ static const char *input_word(struct state *s,const char *pos) {
 
 	while (s->input_begin != s->input_end
 	   &&  s->buffer_end != &s->buffer[sizeof s->buffer]
-           &&  isalnum(*s->input_begin)) {
+           &&  isalnum((int)*s->input_begin)) {
 		if (s->buffer == s->buffer_end
-		|| !isdigit(s->buffer_end[-1]) || !isalpha(*s->input_begin))
+		|| !isdigit((int)s->buffer_end[-1]) || !isalpha((int)*s->input_begin))
 			*s->buffer_end++ = *s->input_begin;
 		++s->input_begin;
 	}
